+++ NEWS.rst
@@ -1,3 +1,12 @@
+Version 1.6.0, 22 January 2020
+------------------------------
+
+New features:
+
+* Support for returning a consolidated output from oio_rest which collapses
+  'virkning' intervals to the smallest amount
+
+
 Version 1.5.0, 04 December 2019
 -------------------------------